:refactor: nextauth
